src/ingestion/requirements.txt
# Satellite Data Processing Dependencies
# Phase 3: Processing Pipeline

# STAC catalog queries
pystac-client>=0.7.0

# Planetary Computer integration
planetary-computer>=1.0.0

# Raster loading and processing
odc-stac>=0.3.0
rasterio>=1.3.0
rioxarray>=0.13.0

# Array operations
xarray>=2023.0.0
numpy>=1.23.0

# Geospatial operations
geopandas>=0.13.0
shapely>=2.0.0

# Image visualization (optional, for debugging)
matplotlib>=3.7.0

# HTTP requests (for Planet API)
requests>=2.28.0

# Date parsing
python-dateutil>=2.8.0

# Scientific image processing
scipy>=1.10.0

# Turf.js alternative for area calculation (if needed in Python)
# pyproj>=3.4.0
